+++ src/openforms/forms/tests/variables/test_variables_in_logic_trigger.py
@@ -61,9 +61,7 @@ class VariablesInLogicBulkAPITests(APITestCase):
                 },
                 "actions": [
                     {
-                        "action": {
-                            "type": "disable-next",
-                        }
+                        "action": {"type": "set-registration-backend", "value": "foo"},
                     }
                 ],
             }
@@ -93,9 +91,7 @@ class VariablesInLogicBulkAPITests(APITestCase):
                 },
                 "actions": [
                     {
-                        "action": {
-                            "type": "disable-next",
-                        }
+                        "action": {"type": "set-registration-backend", "value": "foo"},
                     }
                 ],
             }
@@ -210,9 +206,7 @@ class VariablesInLogicBulkAPITests(APITestCase):
                 },
                 "actions": [
                     {
-                        "action": {
-                            "type": "disable-next",
-                        }
+                        "action": {"type": "set-registration-backend", "value": "foo"},
                     }
                 ],
             }