+++ src/openforms/variables/service.py
@@ -2,6 +2,8 @@
 Public Python API to access (static) form variables.
 """
 
+from collections.abc import Collection
+
 from openforms.forms.models import FormVariable
 from openforms.plugins.registry import BaseRegistry
 from openforms.submissions.models import Submission
@@ -10,7 +12,7 @@ from .base import BaseStaticVariable
 from .registry import register_static_variable as static_variables_registry
 from .utils import get_variables_for_context
 
-__all__ = ["get_static_variables", "get_variables_for_context"]
+__all__ = ["get_static_variables", "get_variables_for_context", "resolve_key"]
 
 
 type VariablesRegistry = BaseRegistry[BaseStaticVariable]
@@ -48,3 +50,35 @@ def get_static_variables(
             and registered_variable.exclude_from_confirmation_email
         )
     ]
+
+
+def resolve_key(input_key: str, all_form_variable_keys: Collection[str]) -> str | None:
+    """
+    Resolve a (nested) key to its corresponding form variable key.
+
+    Submission data of container-type components (e.g. editgrid, selectboxes, etc.)
+    can be accessed with dot notation in a ``FormioData`` instance, but also in JSON
+    logic. For example, key "selectboxes.a" with data ``{"a": "A", "b": "B"}``. This
+    routine resolves the form variable key that corresponds to this data access key.
+
+    :param input_key: The key to resolve.
+    :param all_form_variable_keys: Collection of form variable keys to resolve for.
+    :return: The resolved form variable key, or ``None`` if not resolved.
+    """
+    # There is a variable with this exact key, it is a valid reference.
+    if input_key in all_form_variable_keys:
+        return input_key
+
+    # Process nested paths (editgrid, selectboxes, partners, children). Note that this
+    # doesn't include other nested fields anymore, e.g. a textfield component with key
+    # "foo.bar" will have already been resolved. We process all slices, as these keys
+    # could also include dots.
+    parts = input_key.split(".")
+    for i in range(1, len(parts)):
+        if (key := ".".join(parts[:i])) in all_form_variable_keys:
+            return key
+
+    # If none of the slices exist, we cannot resolve the complete key, so we just
+    # return `None`. Note that the digest email should notify the user of invalid
+    # logic rules.
+    return None